Reference

Index

CUTEst.CUTEstModel

CUTEst.build_libsif

CUTEst.ccf

CUTEst.ccfg

CUTEst.ccfsg

CUTEst.cchprods

CUTEst.cchprodsp

CUTEst.ccifg

CUTEst.ccifsg

CUTEst.cconst

CUTEst.cdh

CUTEst.cdhc

CUTEst.cdimchp

CUTEst.cdimen

CUTEst.cdimse

CUTEst.cdimsh

CUTEst.cdimsj

CUTEst.ceh

CUTEst.cfn

CUTEst.cgr

CUTEst.cgrdh

CUTEst.chcprod

CUTEst.chprod

CUTEst.cidh

CUTEst.cifn

CUTEst.cigr

CUTEst.cisgr

CUTEst.cish

CUTEst.cjprod

CUTEst.clear_libsif

CUTEst.clfg

CUTEst.cnames

CUTEst.cofg

CUTEst.cofsg

CUTEst.connames

CUTEst.cons_coord

CUTEst.cons_coord!

CUTEst.consjac

CUTEst.create_class

CUTEst.creport

CUTEst.csetup

CUTEst.csgr

CUTEst.csgreh

CUTEst.csgrp

CUTEst.csgrsh

CUTEst.csgrshp

CUTEst.csh

CUTEst.cshc

CUTEst.cshcprod

CUTEst.cshp

CUTEst.cshprod

CUTEst.csjp

CUTEst.csjprod

CUTEst.cstats

CUTEst.cterminate

CUTEst.cvartype

CUTEst.fclose

CUTEst.fopen

CUTEst.list_sif_problems

CUTEst.manage_libsif

CUTEst.pname

CUTEst.probname

CUTEst.select

CUTEst.set_mastsif

CUTEst.sifdecoder

CUTEst.ubandh

CUTEst.udh

CUTEst.udimen

CUTEst.udimse

CUTEst.udimsh

CUTEst.ueh

CUTEst.ufn

CUTEst.ugr

CUTEst.ugrdh

CUTEst.ugreh

CUTEst.ugrsh

CUTEst.uhprod

CUTEst.unames

CUTEst.uofg

CUTEst.ureport

CUTEst.usetup

CUTEst.ush

CUTEst.ushp

CUTEst.ushprod

CUTEst.uterminate

CUTEst.uvartype

CUTEst.varnames

CUTEst.create_class — Function

create_class()

Creates the file classf.json, running each problem in $MASTSIF/CLASSF.DB and extracting the necessary information. It should be left alone, unless you think it is not updated. If you do, please open an issue at https://github.com/JuliaSmoothOptimizers/CUTEst.jl

CUTEst.select — Function

select(;min_var=1, max_var=Inf, min_con=0, max_con=Inf,
        objtype=*, contype=*,
        only_free_var=false, only_bnd_var=false,
        only_linear_con=false, only_nonlinear_con=false,
        only_equ_con=false, only_ineq_con=false,
        custom_filter=*)

Returns a subset of the CUTEst problems using the classification file classf.json. This file is export together with the package, so if you have an old CUTEst installation, it can lead to inconsistencies.

min_var and max_var set the number of variables in the problem;

min_con and max_con set the number of constraints in the problem

(e.g., use max_con=0 for unconstrained or min_con=1 for constrained)

only_* flags are self-explaining. Note that they appear in conflicting

pairs. Both can be false, but only one can be true.

objtype is the classification of the objective function according to the

MASTSIF classification file. It can be a number, a symbol, a string, or an array of those.

1, :none or "none" means there is no objective function;
2, :constant or "constant" means the objective function is a constant;
3, :linear or "linear" means the objective function is a linear functional;
4, :quadratic or "quadratic" means the objective function is quadratic;
5, :sum_of_squares or "sum_of_squares" means the objective function is a sum of squares
6, :other or "other" means the objective function is none of the above.

contype is the classification of the constraints according to the same MASTSIF classification file.

1, :unc or "unc" means there are no constraints at all; 2, :fixedvars or "fixedvars" means the only constraints are fixed variables; 3, :bounds or "bounds" means the only constraints are bounded variables; 4, :network or "network" means the constraints represent the adjacency matrix of a (linear) network; 5, :linear or "linear" means the constraints are linear; 6, :quadratic or "quadratic" means the constraints are quadratic; 7, :other or "other" means the constraints are more general.

custom_filter is a function to be applied to the problem data, which is a dict with the following fields:

"objtype" - String one of the above objective function types "contype" - String one of the above constraint types "regular" - Bool whether the problem is regular or not "derivativeorder" - Int order of the highest derivative available "origin" - String origin of the problem: "academic", "modelling" or "real" "hasintervalvar" - Bool whether it has interval variables "variables" - Dict with the following fields "canchoose" - Bool whether you can change the number of variables via parameters "number" - Int the number of variables (if can_choose, the default) "fixed" - Int the number of fixed variables "free" - Int the number of free variables "boundedbelow" - Int the number of variables bounded only from below "boundedabove" - Int the number of variables bounded only from above "boundedboth" - Int the number of variables bounded from below and above "constraints" - Dict with the following fields "canchoose" - Bool whether you can change the number of constraints via parameters "number" - Int the number of constraints (if can_choose, the default) "equality" - Int the number of equality constraints "ineqbelow" - Int the number of inequalities of the form c(x) ≧ cl "ineqabove" - Int the number of inequalities of the form c(x) ≦ cu "ineq_both" - Int the number of inequalities of the form cl ≦ c(x) ≦ cu "linear" - Int the number of linear constraints "nonlinear" - Int the number of nonlinear constraints

For instance, if you'd like to choose only problems with fixed number of variables, you can pass

custom_filter=x->x["variables"]["can_choose"]==false

CUTEst.cons_coord — Function

cons_coord(nlp, x)

Computes the constraint vector and the Jacobian in coordinate format. Usage:

c, jrow, jcol, jval = cons_coord(nlp, x)

nlp: [IN] CUTEstModel

x: [IN] Vector{T}

c: [OUT] Vector{T}

jrow: [OUT] Vector{Cint}

jcol: [OUT] Vector{Cint}

jval: [OUT] Vector{T}

CUTEst.cons_coord! — Function

cons_coord!(nlp, x, c, rows, cols, vals)

Computes the constraint vector and the Jacobian in coordinate format. Usage:

cons_coord!(nlp, x, c, rows, cols, vals)

nlp: [IN] CUTEstModel

x: [IN] Vector{T}

c: [OUT] Vector{T}

jrow: [OUT] Vector{Cint}

jcol: [OUT] Vector{Cint}

jval: [OUT] Vector{T}

CUTEst.consjac — Function

consjac(nlp, x)

Computes the constraint vector and the Jacobian in internal sparse format. Usage:

c, J = consjac(nlp, x)

nlp: [IN] CUTEstModel

x: [IN] Vector{T}

c: [OUT] Vector{T}

J: [OUT] Base.SparseMatrix.SparseMatrixCSC{T,Cint}