Reference

Index

CUTEst.CUTEstModel

CUTEst.build_classification

CUTEst.build_libsif

CUTEst.ccf

CUTEst.ccfg

CUTEst.ccfsg

CUTEst.cchprods

CUTEst.cchprodsp

CUTEst.ccifg

CUTEst.ccifsg

CUTEst.cconst

CUTEst.cdh

CUTEst.cdhc

CUTEst.cdimchp

CUTEst.cdimen

CUTEst.cdimse

CUTEst.cdimsh

CUTEst.cdimsj

CUTEst.ceh

CUTEst.cfn

CUTEst.cgr

CUTEst.cgrdh

CUTEst.chcprod

CUTEst.chjprod

CUTEst.chprod

CUTEst.chsprod

CUTEst.cidh

CUTEst.cifn

CUTEst.cigr

CUTEst.cisgr

CUTEst.cish

CUTEst.cjprod

CUTEst.classification

CUTEst.clear_libsif

CUTEst.clfg

CUTEst.cnames

CUTEst.cofg

CUTEst.cofsg

CUTEst.cohprods

CUTEst.connames

CUTEst.cons_coord

CUTEst.cons_coord!

CUTEst.consjac

CUTEst.creport

CUTEst.csetup

CUTEst.csgr

CUTEst.csgreh

CUTEst.csgrp

CUTEst.csgrsh

CUTEst.csgrshp

CUTEst.csh

CUTEst.cshc

CUTEst.cshcprod

CUTEst.cshp

CUTEst.cshprod

CUTEst.csjp

CUTEst.csjprod

CUTEst.cstats

CUTEst.cterminate

CUTEst.cvartype

CUTEst.fclose

CUTEst.fopen

CUTEst.generator_sif_problems

CUTEst.list_sif_problems

CUTEst.manage_libsif

CUTEst.pname

CUTEst.probname

CUTEst.select_sif_problems

CUTEst.set_mastsif

CUTEst.sifdecoder

CUTEst.ubandh

CUTEst.udh

CUTEst.udimen

CUTEst.udimse

CUTEst.udimsh

CUTEst.ueh

CUTEst.ufn

CUTEst.ugr

CUTEst.ugrdh

CUTEst.ugreh

CUTEst.ugrsh

CUTEst.uhprod

CUTEst.unames

CUTEst.uofg

CUTEst.ureport

CUTEst.usetup

CUTEst.ush

CUTEst.ushp

CUTEst.ushprod

CUTEst.uterminate

CUTEst.uvartype

CUTEst.varnames

CUTEst.cons_coord — Function

cons_coord(nlp, x)

Computes the constraint vector and the Jacobian in coordinate format. Usage:

c, jrow, jcol, jval = cons_coord(nlp, x)

nlp: [IN] CUTEstModel

x: [IN] Vector{T}

c: [OUT] Vector{T}

jrow: [OUT] Vector{Cint}

jcol: [OUT] Vector{Cint}

jval: [OUT] Vector{T}

CUTEst.cons_coord! — Function

cons_coord!(nlp, x, c, rows, cols, vals)

Computes the constraint vector and the Jacobian in coordinate format. Usage:

cons_coord!(nlp, x, c, rows, cols, vals)

nlp: [IN] CUTEstModel

x: [IN] Vector{T}

c: [OUT] Vector{T}

jrow: [OUT] Vector{Cint}

jcol: [OUT] Vector{Cint}

jval: [OUT] Vector{T}

CUTEst.consjac — Function

consjac(nlp, x)

Computes the constraint vector and the Jacobian in internal sparse format. Usage:

c, J = consjac(nlp, x)

nlp: [IN] CUTEstModel

x: [IN] Vector{T}

c: [OUT] Vector{T}

J: [OUT] Base.SparseMatrix.SparseMatrixCSC{T,Cint}